+++ doc/release-notes.md
@@ -17,6 +17,7 @@
 - Add settings in basicswap.json to set Monero rpc timeouts
   - `rpctimeout`, `walletrpctimeout` and `walletrpctimeoutlong` in the Monero section of basicswap.json.
   - `wallet_update_timeout` in basicswap.json to set how long the wallet ui page waits for an rpc response.
+- ui: Renamed unconfirmed balance to pending and include immature balance in pending.
 
 
 0.12.6